Build storage priority: higher-priority boxes pull from lower ones
Phase 2 of the storage system. A box pulls an item from strictly-lower-priority storage (plus loose piles and harvestables), never from same or higher priority, so items flow monotonically up the Low/Medium/High/Critical ladder and A->B->A cycles are structurally impossible without a reservation system. The pull reuses the existing umbrella-Haul goal and deposit leg; the only new action is a Withdraw-from-container pickup (ActionType::Withdraw + Task.haulSourceStorageId), which seats two-hand goods into the hands like every other collection path. Quantity is clamped by the source's own min, the destination's max, and per-trip carry.

Storage priority orders stocking within tier 6 via a score bias, so a Critical box's restock outranks a Low box's but can never preempt a need or an active work order.

Also fixes a latent bug: the capability bitmask was uint8_t sized for 7 capabilities, so Storage (bit 8) and Craftable (bit 7) overflowed and were never recorded in colonist Memory. Widened to uint16_t / 9. Without it colonists could never know a storage box, so the discovery-gated pull would never fire.

890 engine-tests green (+11).

+++ libs/engine/ecs/systems/VisionSystem.cpp
@@ -81,7 +81,7 @@ namespace ecs {
 		auto& registry = engine::assets::AssetRegistry::Get();
 
 		// Create capability mask for shore (Drinkable - colonists drink AT the shore)
-		m_shoreTileCapabilityMask = (1 << static_cast<uint8_t>(engine::assets::CapabilityType::Drinkable));
+		m_shoreTileCapabilityMask = static_cast<uint16_t>(1U << static_cast<size_t>(engine::assets::CapabilityType::Drinkable));
 
 		// Register synthetic shore tile definition
 		m_shoreTileDefNameId = registry.registerSyntheticDefinition(kShoreTileDefName, m_shoreTileCapabilityMask);
@@ -345,7 +345,7 @@ namespace ecs {
 							if (defNameId == 0) {
 								continue;
 							}
-							uint8_t capabilityMask = registry.getCapabilityMask(defNameId);
+							uint16_t capabilityMask = registry.getCapabilityMask(defNameId);
 
 							// Remember in colonist's memory (returns true only for NEW discoveries)
 							bool isNewDiscovery = memory.rememberWorldEntity(placedEntity->position, defNameId, capabilityMask);
@@ -383,7 +383,7 @@ namespace ecs {
 					// Get defNameId and capability mask from registry
 					uint32_t defNameId = registry.getDefNameId(appearance.defName);
 					if (defNameId != 0) {
-						uint8_t capabilityMask = registry.getCapabilityMask(defNameId);
+						uint16_t capabilityMask = registry.getCapabilityMask(defNameId);
 
 						// Remember in colonist's memory (returns true only for NEW discoveries)
 						bool isNewDiscovery = memory.rememberWorldEntity(otherPos.value, defNameId, capabilityMask);
